online chatting

+++ src/app/api/posts/[postId]/likes/route.ts
@@ -14,7 +14,6 @@ export async function GET(
       return Response.json({ error: "未授权" }, { status: 401 });
     }
 
-
     const post = await prisma.post.findUnique({
       where: { id: postId },
       select: {
@@ -62,38 +61,44 @@ export async function POST(
     }
 
     const post = await prisma.post.findUnique({
-      where: {id: postId},
+      where: { id: postId },
       select: {
-        userId: true
-      }
-    })
+        userId: true,
+      },
+    });
 
-    if(!post) {
+    if (!post) {
       return Response.json({ error: "帖子不存在" }, { status: 404 });
     }
 
-    await prisma.like.upsert({
-      where: {
-        userId_postId: {
+    await prisma.$transaction([
+      prisma.like.upsert({
+        where: {
+          userId_postId: {
+            userId: loggedInUser.id,
+            postId,
+          },
+        },
+        create: {
           userId: loggedInUser.id,
           postId,
         },
-      },
-      create: {
-        userId: loggedInUser.id,
-        postId,
-      },
-      update: {},
-    });
+        update: {},
+      }),
+      ...(loggedInUser.id !== post.userId
+        ? [
+            prisma.notification.create({
+              data: {
+                issuerId: loggedInUser.id,
+                recipientId: post.userId,
+                postId,
+                type: "LIKE",
+              },
+            }),
+          ]
+        : []),
+    ]);
 
-    await prisma.notification.create({
-      data: {
-        issuserId: loggedInUser.id,
-        recipientId: post.userId,
-        postId,
-        type: "LIKE"
-      }
-    })
     return new Response();
   } catch (error) {
     console.error(error);
@@ -113,12 +118,33 @@ export async function DELETE(
       return Response.json({ error: "未授权" }, { status: 401 });
     }
 
-    await prisma.like.deleteMany({
-      where: {
-        userId: loggedInUser.id,
-        postId,
+    const post = await prisma.post.findUnique({
+      where: { id: postId },
+      select: {
+        userId: true,
       },
     });
+
+    if (!post) {
+      return Response.json({ error: "帖子不存在" }, { status: 404 });
+    }
+
+    await prisma.$transaction([
+      prisma.like.deleteMany({
+        where: {
+          userId: loggedInUser.id,
+          postId,
+        },
+      }),
+      prisma.notification.deleteMany({
+        where: {
+          issuerId: loggedInUser.id,
+          recipientId: post.userId,
+          postId,
+          type: "LIKE",
+        },
+      }),
+    ]);
 
     return new Response();
   } catch (error) {
